requirements.txt
flask
pypdf
python-docx
openpyxl
pillow
pywin32
easyocr
docx2pdf
